Implemented the PLAIN authentication (binary only). This brings a breaking change in the ServerPool, so from now on I will not make any effort to make the code .NET 2.0 compatible (it's been 2006 and 2 major releases ago!). AssemblyInfo version number will be bumped to 2.x later.
The MemcachedClient class should be instantiated the least possible times, since its constructor is quite expensive. Additionally, every time it is created it will connect to the Memcached pool and create a lot of presistent connections.

Previously people did not pay attention to this and created MemcachedClients every time they needed to connect to Memcached. Moreover the client was not Disposes either, so they ended up with several thousands of connections to the servers.

A hack was introduced to deal with this: the NodeFactory shared MemcachedNodes amongst the client instances, thus unlimited number of client could be created, the socket pool connection limits were maintained.

With the introduction of authentication this is no longer possible (without hacking the code and/or rewriting most of the socket pool), so NodeFactory had to go.

The upcoming documentation, or at least the release notes should stress, that you create only one MemcachedClient per pool/configuration, and use it in through your application's lifetime. (Or Dispose it as soon as it is not needed, but this will introduce performance problems.)