Try to detect when a node is diconnected by cutting the LAN cable. (As in 'the socket is not closed properly'.)
The MessageStreamListener will create a timer which tries to connect to the pools start url (/ppols/default) in every 10 seconds. If it fails then the node is down and it moves over to the next.

The problem comes from the fact that the 'pool streaming uri' is a comet style uri whihc must be read indefinitely. However the nodes only send data very rarey -- when the configuration change. So when a node is disconnected forcefully (by cutting the LAN cabel or disconnecting its VM) we have no way to figure out that no data is coming because the configuration did not change, or because the server is not there.

It'd be more easier if the server'd send an "I am alive" message every 30 seconds or so.