NCBC-228: Frequent pooled socket errors during rebalance
It appears that there are concurrency issues when the server
is sending frequent config updates, as is the case during
a rebalance.  Config updates trigger the client to destroy its
socket pool after first creating a new one.

This change briefly delays the destruction of the old pool
so that any in progress operations may complete.